Major overhaul of tcp implementation
Now using the same crypto stuff as the zeromq implementation
